refactor: updated migration files for dev
