+++ alembic/versions/16ce1dc047b0_added_provider_model.py
@@ -1,0 +1,87 @@
+"""Added Provider Model
+
+Revision ID: 16ce1dc047b0
+Revises: b6c2eb9ea3bc
+Create Date: 2025-05-22 05:04:32.781409
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = '16ce1dc047b0'
+down_revision: Union[str, None] = 'b6c2eb9ea3bc'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    sa.Enum('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', 'RERANK', 'MODERATION', name='modelendpointenum').create(op.get_bind())
+    sa.Enum('TEXT_INPUT', 'TEXT_OUTPUT', 'IMAGE_INPUT', 'IMAGE_OUTPUT', 'AUDIO_INPUT', 'AUDIO_OUTPUT', name='modalityenum').create(op.get_bind())
+    op.create_table('provider',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('provider_type', sa.String(), nullable=False),
+    sa.Column('icon', sa.String(), nullable=False),
+    sa.Column('description', sa.String(), nullable=False),
+    sa.Column('credentials', postgresql.JSONB(astext_type=sa.Text()), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('modified_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('provider_type')
+    )
+    op.create_table('model_info',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('uri', sa.String(), nullable=False),
+    sa.Column('modality', postgresql.ARRAY(postgresql.ENUM('TEXT_INPUT', 'TEXT_OUTPUT', 'IMAGE_INPUT', 'IMAGE_OUTPUT', 'AUDIO_INPUT', 'AUDIO_OUTPUT', name='modalityenum', create_type=False)), nullable=True),
+    sa.Column('input_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('output_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('cache_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('search_context_cost_per_query', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('tokens', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('rate_limits', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('media_limits', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('features', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    sa.Column('provider_id', sa.UUID(), nullable=False),
+    sa.Column('deprecation_date', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('endpoints', postgresql.ARRAY(postgresql.ENUM('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', 'RERANK', 'MODERATION', name='modelendpointenum', create_type=False)), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('modified_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['provider_id'], ['provider.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('uri')
+    )
+    op.create_table('engine_version_model_info',
+    sa.Column('engine_version_id', sa.UUID(), nullable=False),
+    sa.Column('model_info_id', sa.UUID(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['engine_version_id'], ['engine_version.id'], ),
+    sa.ForeignKeyConstraint(['model_info_id'], ['model_info.id'], ),
+    sa.PrimaryKeyConstraint('engine_version_id', 'model_info_id')
+    )
+    op.create_table('engine_version_provider',
+    sa.Column('engine_version_id', sa.UUID(), nullable=False),
+    sa.Column('provider_id', sa.UUID(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['engine_version_id'], ['engine_version.id'], ),
+    sa.ForeignKeyConstraint(['provider_id'], ['provider.id'], ),
+    sa.PrimaryKeyConstraint('engine_version_id', 'provider_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('engine_version_provider')
+    op.drop_table('engine_version_model_info')
+    op.drop_table('model_info')
+    op.drop_table('provider')
+    sa.Enum('TEXT_INPUT', 'TEXT_OUTPUT', 'IMAGE_INPUT', 'IMAGE_OUTPUT', 'AUDIO_INPUT', 'AUDIO_OUTPUT', name='modalityenum').drop(op.get_bind())
+    sa.Enum('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', 'RERANK', 'MODERATION', name='modelendpointenum').drop(op.get_bind())
+    # ### end Alembic commands ###
