--- alembic/versions/1d253a84e955_added_provider_model_table.py
@@ -1,53 +1,0 @@
-"""Added Provider Model Table
-
-Revision ID: 1d253a84e955
-Revises: b6c2eb9ea3bc
-Create Date: 2025-05-14 13:17:17.252151
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '1d253a84e955'
-down_revision: Union[str, None] = 'b6c2eb9ea3bc'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('provider',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('provider_type', sa.String(), nullable=False),
-    sa.Column('icon', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('modified_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('provider_type')
-    )
-    op.create_table('model_info',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('uri', sa.String(), nullable=False),
-    sa.Column('modality', postgresql.ARRAY(sa.Enum('TEXT_INPUT', 'TEXT_OUTPUT', 'IMAGE_INPUT', 'IMAGE_OUTPUT', 'AUDIO_INPUT', 'AUDIO_OUTPUT', name='modalityenum')), nullable=True),
-    sa.Column('config', postgresql.JSONB(astext_type=sa.Text()), nullable=False),
-    sa.Column('provider_id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('modified_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['provider_id'], ['provider.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('uri')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('model_info')
-    op.drop_table('provider')
-    # ### end Alembic commands ###
