--- alembic/versions/3d423f56d107_created_engine_provider_association.py
@@ -1,38 +1,0 @@
-"""Created Engine Provider Association
-
-Revision ID: 3d423f56d107
-Revises: 1d253a84e955
-Create Date: 2025-05-15 00:15:57.394842
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3d423f56d107'
-down_revision: Union[str, None] = '1d253a84e955'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('engine_version_provider',
-    sa.Column('engine_version_id', sa.UUID(), nullable=False),
-    sa.Column('provider_id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['engine_version_id'], ['engine_version.id'], ),
-    sa.ForeignKeyConstraint(['provider_id'], ['provider.id'], ),
-    sa.PrimaryKeyConstraint('engine_version_id', 'provider_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('engine_version_provider')
-    # ### end Alembic commands ###
