--- alembic/versions/405cce2ba1cb_added_model_endpoints_enum.py
@@ -1,43 +1,0 @@
-"""Added Model Endpoints Enum
-
-Revision ID: 405cce2ba1cb
-Revises: 9d37bee05ee6
-Create Date: 2025-05-18 06:42:32.322242
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from alembic_postgresql_enum import ColumnType
-from alembic_postgresql_enum import TableReference
-
-# revision identifiers, used by Alembic.
-revision: str = '405cce2ba1cb'
-down_revision: Union[str, None] = '9d37bee05ee6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema='public',
-        enum_name='modelendpointenum',
-        new_values=['CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', 'RERANK', 'MODERATION'],
-        affected_columns=[TableReference(table_schema='public', table_name='model_info', column_name='endpoints', column_type=ColumnType.ARRAY)],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema='public',
-        enum_name='modelendpointenum',
-        new_values=['CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE'],
-        affected_columns=[TableReference(table_schema='public', table_name='model_info', column_name='endpoints', column_type=ColumnType.ARRAY)],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
