--- alembic/versions/9d37bee05ee6_added_model_info_schema.py
@@ -1,62 +1,0 @@
-"""Added Model Info Schema
-
-Revision ID: 9d37bee05ee6
-Revises: 3d423f56d107
-Create Date: 2025-05-16 14:02:00.373204
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '9d37bee05ee6'
-down_revision: Union[str, None] = '3d423f56d107'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    sa.Enum('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', name='modelendpointenum').create(op.get_bind())
-    op.create_table('engine_version_model_info',
-    sa.Column('engine_version_id', sa.UUID(), nullable=False),
-    sa.Column('model_info_id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['engine_version_id'], ['engine_version.id'], ),
-    sa.ForeignKeyConstraint(['model_info_id'], ['model_info.id'], ),
-    sa.PrimaryKeyConstraint('engine_version_id', 'model_info_id')
-    )
-    op.add_column('model_info', sa.Column('input_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('output_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('cache_cost', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('search_context_cost_per_query', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('tokens', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('rate_limits', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('media_limits', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('features', postgresql.JSONB(astext_type=sa.Text()), nullable=True))
-    op.add_column('model_info', sa.Column('deprecation_date', sa.DateTime(timezone=True), nullable=True))
-    op.add_column('model_info', sa.Column('endpoints', postgresql.ARRAY(postgresql.ENUM('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', name='modelendpointenum', create_type=False)), nullable=True))
-    op.drop_column('model_info', 'config')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('model_info', sa.Column('config', postgresql.JSONB(astext_type=sa.Text()), autoincrement=False, nullable=False))
-    op.drop_column('model_info', 'endpoints')
-    op.drop_column('model_info', 'deprecation_date')
-    op.drop_column('model_info', 'features')
-    op.drop_column('model_info', 'media_limits')
-    op.drop_column('model_info', 'rate_limits')
-    op.drop_column('model_info', 'tokens')
-    op.drop_column('model_info', 'search_context_cost_per_query')
-    op.drop_column('model_info', 'cache_cost')
-    op.drop_column('model_info', 'output_cost')
-    op.drop_column('model_info', 'input_cost')
-    op.drop_table('engine_version_model_info')
-    sa.Enum('CHAT', 'COMPLETION', 'IMAGE_GENERATION', 'AUDIO_TRANSCRIPTION', 'AUDIO_SPEECH', 'EMBEDDING', 'BATCH', 'RESPONSE', name='modelendpointenum').drop(op.get_bind())
-    # ### end Alembic commands ###
