--- alembic/versions/c3d0ed7510fd_added_credentials_column_in_provider.py
@@ -1,30 +1,0 @@
-"""Added Credentials Column In Provider
-
-Revision ID: c3d0ed7510fd
-Revises: 405cce2ba1cb
-Create Date: 2025-05-19 00:43:48.425499
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'c3d0ed7510fd'
-down_revision: Union[str, None] = '405cce2ba1cb'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('provider', sa.Column('credentials', postgresql.JSONB(astext_type=sa.Text()), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('provider', 'credentials')
-    # ### end Alembic commands ###
